+++ src/Hangfire.PostgreSql/IPersistentJobQueueMonitoringApi.cs
@@ -26,8 +26,8 @@ namespace Hangfire.PostgreSql
     public interface IPersistentJobQueueMonitoringApi
     {
         IEnumerable<string> GetQueues();
-        IEnumerable<int> GetEnqueuedJobIds(string queue, int from, int perPage);
-        IEnumerable<int> GetFetchedJobIds(string queue, int from, int perPage);
+        IEnumerable<long> GetEnqueuedJobIds(string queue, int from, int perPage);
+        IEnumerable<long> GetFetchedJobIds(string queue, int from, int perPage);
         EnqueuedAndFetchedCountDto GetEnqueuedAndFetchedCount(string queue);
     }
 }
