+++ src/Hangfire.PostgreSql/PostgreSqlJobQueue.cs
@@ -239,8 +239,8 @@ VALUES (@jobId, @queue);
 		[UsedImplicitly(ImplicitUseTargetFlags.WithMembers)]
 		private class FetchedJob
 		{
-			public int Id { get; set; }
-			public int JobId { get; set; }
+			public long Id { get; set; }
+			public long JobId { get; set; }
 			public string Queue { get; set; }
 			public DateTime? FetchedAt { get; set; }
 			public int UpdateCount { get; set; }
